+++ changelog.md
@@ -18,6 +18,8 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
   with, rather than in place of, `semigroupLaws` and `monoidLaws`. This is a breaking
   change.
 - Fix the right distribution law for semirings.
+- The function `lawsCheckMany` now terminates with exit code 1 if a
+  test fails.
 ### Added
 - Add `genericLaws` and `generic1Laws`
 - Add property tests for special classes of semigroups. This includes: